requirements.txt
# Phase 1 — Banking BI Copilot CDK dependencies
aws-cdk-lib>=2.130.0
constructs>=10.3.0
aws-cdk.aws-glue-alpha>=2.130.0a0
boto3>=1.34.0
